[breaking] Response.headers is now of type Dict{String, Vector{String}}
In the past, Response.headers was Dict{String, String}. This meant
that when there were headers with duplicate keys, only one of those
headers would be visible to the caller. A common example of this is
the "Set-Cookie" header.

Now we'll use Dict{String, Vector{String}} so that all the headers
will be visible.

fixes #11